+++ release_notes.md
@@ -10,3 +10,4 @@
 - Add Dockerfile for python 3.13 local build environment (#4611)
 - Add support for Python 3.14 (#4668)
 - Implement preview feature to apply host configuration profiles on 'func init' command (#4675)
+- Add updated Durable .NET templates (#4692)