+++ release_notes.md
@@ -1,9 +1,9 @@
-# Azure Functions CLI 4.4.0
+# Azure Functions CLI 4.5.0
 
 #### Host Version
 
 - Host Version: 4.1043.200
-- In-Proc Host Version: 4.41.100 (4.841.100, 4.641.100)
+- In-Proc Host Version: 4.44.100 (4.844.100, 4.644.100)
 
 #### Changes